Add begin_frame() to driver
This adds a begin_frame() method to the VulkanDriver which tries to
acquire the next image from the swapchain, setting up the appropriate
semaphore for when the image is acquired.

It also adds initial per frame data which is stored by the driver for
however many frame contexts are required.