Add IntrusivePtr
This adds an intrusive pointer implementation which can use either the
single- or multi-threaded reference tracker, allows for either
thread-safe or non-thread-safe implementations.

The IntrusivePtr is essentially a shared_ptr, but it stored the ref
count locally, so that the overall type is smaller. It's intention is to
be used as a handle type for resources.