Add Allocator implementation
This adds the Allocator<Primary, Fallback, Locking> class which is a
composable allocator, allocating from Primary (which itself could be a
composed allocator), and then usign Fallback if the allocation from
Primary fails.

It also allows the Locking policy to be specified, so that the
different strategies can be used for different threading requirements.