Fix clean-up in engine and vulkan driver
This also adds a fix for surface creation where the present queue was
assumed to be the same as the graphics queue. This is now checked and
the surface context has it's own present queue, which should be the same
as the graphics queue in most contexts.

It also adds a step during presentation to set the rendering done
semaphore before presenting. This is just temporary, and will be moved
into the submission with the last submitted command buffer.